+++ src/DynamoCore/Models/NoteModel.cs
@@ -33,7 +33,7 @@ namespace Dynamo.Models
         protected override bool UpdateValueCore(UpdateValueParams updateValueParams)
         {
             string name = updateValueParams.PropertyName;
-            string value = updateValueParams.Value;
+            string value = updateValueParams.PropertyValue;
 
             if (name != "Text") 
                 return base.UpdateValueCore(updateValueParams);
