+++ src/DynamoCore/Nodes/CodeBlockNode.cs
@@ -235,7 +235,7 @@ namespace Dynamo.Nodes
         protected override bool UpdateValueCore(UpdateValueParams updateValueParams)
         {
             string name = updateValueParams.PropertyName;
-            string value = updateValueParams.Value;
+            string value = updateValueParams.PropertyValue;
 
             if (name != "Code") 
                 return base.UpdateValueCore(updateValueParams);
