+++ src/DynamoCore/Search/SearchDictionary.cs
@@ -229,7 +229,10 @@ namespace Dynamo.Search
             return query.Trim()
                 .Replace("\\", "\\\\")
                 .Replace(".", "\\.")
-                .Replace("*", "\\*");
+                .Replace("*", "\\*")
+                .Replace("(", "\\(")
+                .Replace(")", "\\)")
+                .Replace("[", "\\[");
         }
 
         private static string[] SplitOnWhiteSpace(string s)
