+++ src/Engine/ProtoCore/Namespace/ElementResolver.cs
@@ -12,13 +12,13 @@ namespace ProtoCore.Namespace
     /// </summary>
     public class ElementResolver
     {
-        private Dictionary<string, KeyValuePair<string, string>> resolutionMap;
+        private readonly IDictionary<string, KeyValuePair<string, string>> resolutionMap;
 
         /// <summary>
         /// Maintains a lookup table of partial class identifiers vs. 
         /// fully qualified class identifier names and assembly name
         /// </summary>
-        public Dictionary<string, KeyValuePair<string, string>> ResolutionMap
+        public IDictionary<string, KeyValuePair<string, string>> ResolutionMap
         {
             get { return resolutionMap; }
         }
