+++ src/Libraries/CoreNodeModels/Formula.cs
@@ -62,7 +62,7 @@ namespace DSCoreNodesUI
         protected override bool UpdateValueCore(UpdateValueParams updateValueParams)
         {
             string name = updateValueParams.PropertyName;
-            string value = updateValueParams.Value;
+            string value = updateValueParams.PropertyValue;
 
             if (name == "FormulaString")
             {
