+++ src/Libraries/CoreNodeModels/Input/DoubleSlider.cs
@@ -55,7 +55,7 @@ namespace Dynamo.Nodes
         protected override bool UpdateValueCore(UpdateValueParams updateValueParams)
         {
             string name = updateValueParams.PropertyName;
-            string value = updateValueParams.Value;
+            string value = updateValueParams.PropertyValue;
 
             switch (name)
             {
