+++ src/Libraries/CoreNodeModels/Input/IntegerSlider.cs
@@ -31,7 +31,7 @@ namespace Dynamo.Nodes
         protected override bool UpdateValueCore(UpdateValueParams updateValueParams)
         {
             string name = updateValueParams.PropertyName;
-            string value = updateValueParams.Value;
+            string value = updateValueParams.PropertyValue;
 
             switch (name)
             {
