+++ src/Libraries/IronPython/PythonNode.cs
@@ -158,7 +158,7 @@ namespace DSIronPythonNode
         protected override bool UpdateValueCore(UpdateValueParams updateValueParams)
         {
             string name = updateValueParams.PropertyName;
-            string value = updateValueParams.Value;
+            string value = updateValueParams.PropertyValue;
 
             if (name == "ScriptContent")
             {
